+++ app/include_upgrade_e2e.go
@@ -2,8 +2,18 @@
 
 package app
 
+import (
+	erc20types "github.com/cosmos/evm/x/erc20/types"
+	feemarkettypes "github.com/cosmos/evm/x/feemarket/types"
+	precisebanktypes "github.com/cosmos/evm/x/precisebank/types"
+	evmtypes "github.com/cosmos/evm/x/vm/types"
+
+	v3 "github.com/babylonlabs-io/babylon/v3/app/upgrades/v3"
+)
+
 // init is used to include v1 upgrade testnet data
 // it is also used for e2e testing
 func init() {
 	IsE2EUpgradeBuildFlag = true
+	v3.StoresToAdd = append(v3.StoresToAdd, erc20types.StoreKey, evmtypes.StoreKey, feemarkettypes.StoreKey, precisebanktypes.StoreKey)
 }